+++ input_json.go
@@ -5,31 +5,33 @@ package trdsql
 // * Array ([{c1: 1}, {c1: 2}, {c1: 3}])
 // * Multiple JSON ({c1: 1}\n {c1: 2}\n {c1: 3}\n)
 
+// Make a table from json and path.
 import (
 	"encoding/json"
 	"errors"
 	"io"
 	"log"
+
+	"github.com/Jeffail/gabs/v2"
 )
 
 // JSONReader provides methods of the Reader interface.
 type JSONReader struct {
 	reader  *json.Decoder
 	preRead []map[string]string
+	path    string
 	names   []string
 	types   []string
-	inArray []interface{}
-	count   int
 }
 
 // NewJSONReader returns JSONReader and error.
 func NewJSONReader(reader io.Reader, opts *ReadOpts) (*JSONReader, error) {
 	r := &JSONReader{}
 	r.reader = json.NewDecoder(reader)
-	var top interface{}
-	already := map[string]bool{}
+	r.path = opts.InPath
+
 	for i := 0; i < opts.InPreRead; i++ {
-		row, names, err := r.readAhead(top, i)
+		jsonParsed, err := gabs.ParseJSONDecoder(r.reader)
 		if err != nil {
 			if !errors.Is(err, io.EOF) {
 				return r, err
@@ -37,15 +39,13 @@ func NewJSONReader(reader io.Reader, opts *ReadOpts) (*JSONReader, error) {
 			debug.Printf(err.Error())
 			return r, nil
 		}
-
-		for k := 0; k < len(names); k++ {
-			if !already[names[k]] {
-				already[names[k]] = true
-				r.names = append(r.names, names[k])
-			}
+		if opts.InPath != "" {
+			jsonParsed = jsonParsed.Path(r.path)
 		}
-
-		r.preRead = append(r.preRead, row)
+		r, err = r.readAhead(jsonParsed.Data())
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	return r, nil
@@ -66,50 +66,101 @@ func (r *JSONReader) Types() ([]string, error) {
 	return r.types, nil
 }
 
-func (r *JSONReader) readAhead(top interface{}, count int) (map[string]string, []string, error) {
-	if r.inArray != nil {
-		if len(r.inArray) > count {
-			r.count++
-			return r.secondLevel(top, r.inArray[count])
+func (r *JSONReader) readAhead(top interface{}) (*JSONReader, error) {
+	switch m := top.(type) {
+	case []interface{}:
+		// []
+		r.preRead = make([]map[string]string, 0, len(m))
+		if r.reader.More() {
+			pre, names, err := etcRow(m)
+			if err != nil {
+				return nil, err
+			}
+
+			r.names = names
+			r.preRead = append(r.preRead, pre)
+			return r, nil
 		}
-		return nil, nil, io.EOF
-	}
 
-	err := r.reader.Decode(&top)
-	if err != nil {
-		return nil, nil, err
-	}
-	return r.topLevel(top)
-}
+		already := map[string]bool{}
+		for _, v := range m {
+			pre, names, err := topLevel(v)
+			if err != nil {
+				return nil, err
+			}
 
-func (r *JSONReader) topLevel(top interface{}) (map[string]string, []string, error) {
-	switch obj := top.(type) {
-	case []interface{}:
-		// [{} or [] or etc...]
-		r.inArray = obj
-		return r.secondLevel(top, r.inArray[0])
-	case map[string]interface{}:
-		// {"a":"b"} object
-		r.inArray = nil
-		return objectRow(obj)
-	}
-	return nil, nil, ErrUnableConvert
-}
+			for k := 0; k < len(names); k++ {
+				if !already[names[k]] {
+					already[names[k]] = true
+					r.names = append(r.names, names[k])
+				}
+			}
 
-// Analyze second when top is array.
-func (r *JSONReader) secondLevel(top interface{}, second interface{}) (map[string]string, []string, error) {
-	switch obj := second.(type) {
-	case map[string]interface{}:
-		// [{}]
-		return objectRow(obj)
-	case []interface{}:
-		// [[]]
-		return etcRow(second)
+			r.preRead = append(r.preRead, pre)
+		}
+		return r, nil
 	default:
-		// ["a","b"]
-		r.inArray = nil
-		return etcRow(top)
+		pre, names, err := topLevel(m)
+		if err != nil {
+			return nil, err
+		}
+		r.names = names
+		r.preRead = append(r.preRead, pre)
 	}
+	return r, nil
+}
+
+func topLevel(top interface{}) (map[string]string, []string, error) {
+	switch obj := top.(type) {
+	case map[string]interface{}:
+		return objectRow(obj)
+	default:
+		return etcRow(obj)
+	}
+}
+
+// PreReadRow is returns only columns that store preread rows.
+// One json (not jsonl) returns all rows with preRead.
+func (r *JSONReader) PreReadRow() [][]interface{} {
+	rows := make([][]interface{}, len(r.preRead))
+	for n, v := range r.preRead {
+		rows[n] = make([]interface{}, len(r.names))
+		for i := range r.names {
+			rows[n][i] = v[r.names[i]]
+		}
+	}
+	return rows
+}
+
+// ReadRow is read the rest of the row.
+// Only jsonl requires ReadRow in json.
+func (r *JSONReader) ReadRow(row []interface{}) ([]interface{}, error) {
+	jsonParsed, err := gabs.ParseJSONDecoder(r.reader)
+	if err != nil {
+		return nil, err
+	}
+
+	if r.path != "" {
+		jsonParsed = jsonParsed.Path(r.path)
+	}
+	jsonRow := jsonParsed.Data()
+
+	return r.rowParse(row, jsonRow), nil
+}
+
+func (r *JSONReader) rowParse(row []interface{}, jsonRow interface{}) []interface{} {
+	switch m := jsonRow.(type) {
+	case map[string]interface{}:
+		for i := range r.names {
+			row[i] = jsonString(m[r.names[i]])
+		}
+	default:
+		for i := range r.names {
+			row[i] = nil
+		}
+		row[0] = jsonString(jsonRow)
+	}
+	return row
 }
 
 func objectRow(obj map[string]interface{}) (map[string]string, []string, error) {
@@ -146,60 +197,4 @@ func jsonString(val interface{}) string {
 	default:
 		return ValString(val)
 	}
-}
-
-// PreReadRow is returns only columns that store preread rows.
-func (r *JSONReader) PreReadRow() [][]interface{} {
-	rows := make([][]interface{}, len(r.preRead))
-	for n, v := range r.preRead {
-		rows[n] = make([]interface{}, len(r.names))
-		for i := range r.names {
-			rows[n][i] = v[r.names[i]]
-		}
-	}
-	return rows
-}
-
-// ReadRow is read the rest of the row.
-func (r *JSONReader) ReadRow(row []interface{}) ([]interface{}, error) {
-	if r.inArray != nil {
-		// [] array
-		r.count++
-		if r.count >= len(r.inArray) {
-			var top interface{}
-			err := r.reader.Decode(&top)
-			if err != nil {
-				return nil, err
-			}
-			r.count = 0
-		}
-		if len(r.inArray) > 0 {
-			row = r.rowParse(row, r.inArray[r.count])
-		}
-		return row, nil
-	}
-
-	// {} object
-	var data interface{}
-	err := r.reader.Decode(&data)
-	if err != nil {
-		return nil, err
-	}
-	row = r.rowParse(row, data)
-	return row, nil
-}
-
-func (r *JSONReader) rowParse(row []interface{}, jsonRow interface{}) []interface{} {
-	switch m := jsonRow.(type) {
-	case map[string]interface{}:
-		for i := range r.names {
-			row[i] = jsonString(m[r.names[i]])
-		}
-	default:
-		for i := range r.names {
-			row[i] = nil
-		}
-		row[0] = jsonString(jsonRow)
-	}
-	return row
 }
